+++ pkg/helm/apply.go
@@ -5,10 +5,9 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/replicatedhq/ekco/pkg/helm/chartfiles"
+	"github.com/replicatedhq/ekco/pkg/helm/charts/rookcephcluster"
 	"helm.sh/helm/v3/pkg/action"
 	"helm.sh/helm/v3/pkg/chart/loader"
-	"helm.sh/helm/v3/pkg/chartutil"
 	"helm.sh/helm/v3/pkg/cli"
 )
 
@@ -32,23 +31,15 @@ func ApplyChart(ctx context.Context, chartBytes io.Reader, values map[string]int
 	act.ReleaseName = releaseName
 	act.Wait = false
 
-	cephClusterValues, err := rookCephClusterConfig()
+	chartValues, err := rookcephcluster.ValuesMap()
 	if err != nil {
 		return fmt.Errorf("unable to get %s chart values: %w", chart.Name(), err)
 	}
 
 	// install chart
-	if _, err := act.RunWithContext(ctx, chart, cephClusterValues); err != nil {
+	if _, err := act.RunWithContext(ctx, chart, chartValues); err != nil {
 		return fmt.Errorf("unable to install %s chart: %w", chart.Name(), err)
 	}
 
 	return nil
-}
-
-func rookCephClusterConfig() (map[string]interface{}, error) {
-	values, err := chartutil.ReadValues(chartfiles.CephClusterChartValuesFile)
-	if err != nil {
-		return nil, fmt.Errorf("Unable to read rook-ceph-cluster values file: %w", err)
-	}
-	return values.AsMap(), nil
 }
