--- pkg/helm/chartfiles/embed.go
@@ -1,9 +1,0 @@
-package chartfiles
-
-import "embed"
-
-//go:embed *.tgz
-var FS embed.FS
-
-//go:embed values.yaml
-var CephClusterChartValuesFile []byte
