+++ pkg/helm/charts/rookcephcluster/rook-ceph-cluster-v1.11.6.tgz
